+++ Client/package.json
@@ -13,6 +13,7 @@
     "@nextui-org/react": "^2.4.8",
     "@reduxjs/toolkit": "^2.3.0",
     "alumni": "file:",
+    "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-dropzone": "^14.2.10",
@@ -34,6 +35,7 @@
     "eslint-plugin-react-refresh": "^0.4.9",
     "globals": "^15.9.0",
     "postcss": "^8.4.47",
+    "react-error-overlay": "^6.0.9",
     "tailwindcss": "^3.4.13",
     "vite": "^5.4.1"
   }
