+++ Client/src/components/Login.jsx
@@ -2,12 +2,47 @@ import {Modal, ModalContent, ModalHeader, ModalBody, ModalFooter, Button, useDis
 import {FaUser } from "react-icons/fa";
 import { BiSolidLock, BiSolidLockOpen } from "react-icons/bi";
 import React,{useState} from 'react'
+import axios from "axios";
+import toast from "react-hot-toast";
 
 const Login = ({isOpen,onClose}) =>  {
 
   const {onOpenChange} = useDisclosure();
   const [isVisible,setIsVisible] = useState(false)
   const toggleVisibility = () => setIsVisible(!isVisible)
+  const apiurl = import.meta.env.VITE_API_URL;
+  const [loginCredentials,setLoginCredentials] = useState({
+    alumni_id:'',
+    password:''
+  })
+  const handleLogin = async () => {
+    try {
+      const resp = await axios.post(`${apiurl}/users/login`, {
+        user_id: loginCredentials.alumni_id,
+        password: loginCredentials.password,
+      });
+      const res = resp.data;
+  
+      console.log(res); // Check the entire response
+  
+      if (res.status !== 'success') {
+        setLoginCredentials((prev) => ({
+          ...prev,
+          password: '',
+        }));
+        toast.error(res.message);
+      } else {
+        toast.success('Login Successful');
+        
+        localStorage.setItem('userDetails', JSON.stringify(res.extras[0]));
+        onClose();
+      }
+    } catch (err) {
+      console.error(err);
+      toast.error('An error occurred during login');
+    }
+  };
+  
   return (
     <>
       <Modal 
@@ -29,6 +64,8 @@ const Login = ({isOpen,onClose}) =>  {
                   label="AlumniId"
                   placeholder="Enter your alumniId"
                   variant="bordered"
+                  value={loginCredentials.alumni_id}
+                  onChange={(e)=>setLoginCredentials((prev)=>({...prev,alumni_id:e.target.value}))}
                 />
                 <Input
                   endContent={
@@ -47,15 +84,10 @@ const Login = ({isOpen,onClose}) =>  {
                   placeholder="Enter your password"
                   type={isVisible ? 'text':'password'}
                   variant="bordered"
+                  value={loginCredentials.password}
+                  onChange={(e)=>setLoginCredentials((prev)=>({...prev,password:e.target.value}))}
                 />
-                <div className="flex py-2 px-1 justify-between">
-                  <Checkbox
-                    classNames={{
-                      label: "text-small",
-                    }}
-                  >
-                    Remember me
-                  </Checkbox>
+                <div className="flex py-2 px-1 justify-end">
                   <Link color="primary" href="#" size="sm">
                     Forgot password?
                   </Link>
@@ -65,7 +97,7 @@ const Login = ({isOpen,onClose}) =>  {
                 <Button color="danger" variant="flat" onPress={onClose}>
                   Close
                 </Button>
-                <Button color="primary" onPress={onClose}>
+                <Button color="primary" onPress={handleLogin}>
                   Sign in
                 </Button>
               </ModalFooter>
