+++ Client/vite.config.js
@@ -2,6 +2,10 @@ import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 // https://vitejs.dev/config/
+
 export default defineConfig({
   plugins: [react()],
+  // define: {
+  //   'process.env': {}
+  // }
 })
